Extract audio utils, trivial comments
